Added schedule button that schedules a Recipe
Once the button is clicked by the user, a date picker dialog opens that
allows the user to select which day they'd like to schedule the recipe
for.  Once a date is selected, another dialog opens and prompts the user
to select which meal (breakfast, lunch, dinner) they'd like to schedule
the meal for.  Once selected, the recipe is scheduled.