Checking recipe exists before diplay in Schedule
Added checks to make sure that a recipe exists in the persistence layer
before displaying it in a meal schdule.  There was a bug where a recipe
was still displayed in the schedulea after being deleted.
New methods were added to the business and persistence layer to make
these checks convenient.
Tests were adjusted to reflect changes in DaySchedule.