This commit is mostly focused on moving towards finishing the editing side of ingredients. Things that are needed are setters for deleting / adding a new ingredient to a recipe. Onclick interactions for delete are dependent on that as well. As of now, editing is done via textBox which could definitely lead to some user-induced errors in testing such as an invalid unit of measure or a negative amount of an ingredient.
Later on, I plan to have the UOMs as a dropdown list to select from, and have the fields auto-update conversions factors.

Additionally, I have yet to add the servings multiplier, may be able to get that done before the next iteration/sprint cycle, but that comes after we get DB work done, cleaned up, and properly integrated.